Add refactoring of the components
