+++ packages/core/ui/App.js
@@ -1,20 +1,16 @@
-import AppBar from '@material-ui/core/AppBar'
-import CircularProgress from '@material-ui/core/CircularProgress'
-import Icon from '@material-ui/core/Icon'
-import IconButton from '@material-ui/core/IconButton'
-import Slide from '@material-ui/core/Slide'
 import { makeStyles } from '@material-ui/core/styles'
-import Toolbar from '@material-ui/core/Toolbar'
-import Typography from '@material-ui/core/Typography'
+
 import { observer, PropTypes } from 'mobx-react'
+import { isAlive } from 'mobx-state-tree'
 import ReactPropTypes from 'prop-types'
 import React, { useEffect } from 'react'
 import { withContentRect } from 'react-measure'
-import { inDevelopment } from '../util'
-import DevTools from './DevTools'
-import Drawer from './Drawer'
 
-const useStyles = makeStyles(theme => ({
+import { inDevelopment } from '../util'
+import DrawerWidget from './DrawerWidget'
+import DevTools from './DevTools'
+
+const useStyles = makeStyles(() => ({
   '@global': {
     html: {
       'font-family': 'Roboto',
@@ -40,89 +36,21 @@ const useStyles = makeStyles(theme => ({
   components: {
     display: 'block',
   },
-  drawerCloseButton: {
-    float: 'right',
-  },
-  drawerToolbar: {
-    paddingLeft: theme.spacing(2),
-  },
-  drawerToolbarCloseButton: {
-    flexGrow: 1,
-  },
-  drawerLoading: {
-    margin: theme.spacing(2),
-  },
 }))
-
-const DrawerWidget = observer(props => {
-  const { session } = props
-  const { visibleDrawerWidget, pluginManager } = session
-  const {
-    LazyReactComponent,
-    HeadingComponent,
-    heading,
-  } = pluginManager.getDrawerWidgetType(visibleDrawerWidget.type)
-  const classes = useStyles()
-
-  return (
-    <Drawer session={session} open={Boolean(session.activeDrawerWidgets.size)}>
-      <Slide direction="left" in>
-        <div className={classes.defaultDrawer}>
-          <AppBar position="static" color="secondary">
-            <Toolbar
-              variant="dense"
-              disableGutters
-              className={classes.drawerToolbar}
-            >
-              <Typography variant="h6" color="inherit">
-                {HeadingComponent ? (
-                  <HeadingComponent model={visibleDrawerWidget} />
-                ) : (
-                  heading || undefined
-                )}
-              </Typography>
-              <div className={classes.drawerToolbarCloseButton} />
-              <IconButton
-                className={classes.drawerCloseButton}
-                color="inherit"
-                aria-label="Close"
-                onClick={() => session.hideDrawerWidget(visibleDrawerWidget)}
-              >
-                <Icon fontSize="small">close</Icon>
-              </IconButton>
-            </Toolbar>
-          </AppBar>
-          <React.Suspense
-            fallback={
-              <CircularProgress
-                disableShrink
-                className={classes.drawerLoading}
-              />
-            }
-          >
-            <LazyReactComponent model={visibleDrawerWidget} session={session} />
-          </React.Suspense>
-        </div>
-      </Slide>
-    </Drawer>
-  )
-})
-
-DrawerWidget.propTypes = {
-  session: PropTypes.observableObject.isRequired,
-}
 
 function App({ contentRect, measureRef, session }) {
   const classes = useStyles()
   const { pluginManager } = session
-
   useEffect(() => {
     if (contentRect.bounds.width) {
-      session.updateWidth(contentRect.bounds.width)
+      if (isAlive(session)) {
+        session.updateWidth(contentRect.bounds.width)
+      }
     }
   }, [session, contentRect])
 
   const { visibleDrawerWidget } = session
+
   return (
     <div ref={measureRef} className={classes.root}>
       <div className={classes.menuBarsAndComponents}>
