+++ packages/core/ui/index.ts
@@ -1,8 +1,10 @@
-export { default as theme } from './theme'
+export { default as useTheme } from './theme'
 export { default as LogoFull } from './LogoFull'
 export { default as App } from './App'
 export { default as FileSelector } from './FileSelector'
 export { default as PrerenderedCanvas } from './PrerenderedCanvas'
 export { default as ResizeHandle } from './ResizeHandle'
 export { default as StartScreen } from './StartScreen'
+export { default as FactoryResetDialog } from './FactoryResetDialog'
 export { default as Tooltip } from './Tooltip'
+export { default as FatalErrorDialog } from './FatalErrorDialog'
